boundary transformations switched to double precision. ADded transformations to offlatice class. Fixed bug in transformations in sossolver where x and y boundaries were used arbitrary. Added images to boundaries, so that stuff like closest distance can be calculated. Changed so that a padding of only 0.5 is used around surface sites for block checks to avoid an infinitely small search for an available deposition site. Added check for 0 paths in dissolution for continuum.

+++ src/kmcsolver/boundary/edge.cpp
@@ -11,7 +11,7 @@ Edge::Edge(const int location,
 
 }
 
-bool Edge::isBlocked(const int xi) const
+bool Edge::isBlocked(const double xi) const
 {
     if (m_orientation == orientations::FIRST)
     {
